New player movement style.
I have changed the movement style to the one we discussed. You can use ZXCV to move the player left by varying amounts and NM,. to move the player right by varying amounts. Rapidly pressing the keys will make the player move faster. The effect of friction also increases with speed, so the faster you press the keys, the higher the equilibrium speed with friction.

I added proper keybindings using the "Project Settings > Input Map" menu.

I also changed how the camera works. It now just faces the same way as the player and the code is simpler. The camera needs to be told where to find the player node - i have done this but if you move the camera or the player it might break. 

I added an arrow underneath the player, you can remove it if you want. I needed it there to debug while I worked on this because the Player Sprite3D doesn't show what direction the player is actually going.