Tweak speed and friction.
I wasn't actually using the friction variable, I fixed that. Reducing the friction and the player speed makes the movement feel smoother. Try also hiding the arrow (Because it visually responds immediately to the change in speed, unlike the player sprite, it makes the movement feel more jumpy).
The changes to all the "*.import" files are because I updated to Godot 3.2.1 (I had 3.1 accidentally on my laptop still.)