Add a timer to input.
Hey, I love all the new art and the NPC runner.

Here is a timer. It is configurable in the KeypressTimer node in the player. It works but I don't think it feels very good... Maybe some visual feedback would improve it, like syncing the player running animation to the running. Or, maybe another solution would be better.