Still stuck
I still can't seem to connect the dots with the Timer and the NPC. The best I've been able to do is erase the physics function completely and put the if statements containing the range of speeds underneath the timer_on_timeout function. That leads to the npc remaining still but it doesn't last 3 seconds as intended.