8 Direction Animation
By replacing frame in line 36 of the Sprite3D script found in the Teammate1 node with the exported variable anim_col the animation is able to run but the sprite is stuck staring at me head on no matter which angle I approach it from. The exact same problem happens if I remove the self from the new script attached to the camera. Any idea how to get the animation moving without compromising the angles I’ve entered in the Sprite 3D script?

Also, enjoy a rough version of the stamina meter. The faster you go the quicker it depletes. If you can get close enough to the npc your stamina is replenished.  It also refills once you cross over the road. Still needs a lot of fine tuning but psyched that the concept works.